+++ Facility.Web/Views/Home/Index.cshtml
@@ -8,7 +8,7 @@
 
 <div class="row">
     <div class="col-xs-12">
-        @using (Html.BeginForm("Index", "Produce", FormMethod.Post))
+        @using (Html.BeginForm("Produce", "Home", FormMethod.Post))
         {
         <button type="submit" class="btn btn-default btn-lg">
             <span class="glyphicon glyphicon-star" aria-hidden="true"></span> Produce chocolate
